Promote  to  and make its lookup a Map
