+++ src/codex/CLAUDE.md
@@ -18,4 +18,5 @@ This directory hosts three tools: **Cardex** (`cards.tsx`), **Skilldex** (`skill
 - **Nil (0) expansion is labelled `Core°` in Cardex, shows no card set at all in Skilldex, and is only *selectable* in Cardex.** `useCardSetFilters.ts` exports `useCardSetFilters` (Cardex) and `useTalentCardSetFilters` (Skilldex) over a shared base: the base remaps index 0 → Core for the *name*, but only Cardex remaps it for *selection*. Talents use expansion 0 as a structural marker (event-only talents, offers, some unavailable talents) and are filtered by their own tree sections, so they must stay visible when Core is unticked — `isIndexSelected(0)` must stay false there. Never put index 0 in the shared `indexMap`/`indexToValueMap`: that would make it selectable for talents too. Cardex renders the Core name plus a `°` span; Skilldex excludes index 0 in **both** `TalentTree`'s `shouldShowCardSet` (dimensions) and its `getCardSetName` (render) — those two must agree, or nil nodes reserve height for a row they never draw.
 - **Adding or removing a filter option does not need a `codexFilterStore.ts` version bump.** `createFilterHook` merges the cached blob by iterating the *cache* but gating on `key in defaultFilters`, so stale keys are dropped and missing keys fall back to the default. Removing an option silently ignores the cached value; the key is purged from localStorage on the next debounced write. Only bump the version if the *meaning* or value type of an existing key changes.
 - Cards/talents come from Supabase (24h localStorage cache; versioned keys in `codexCardsStore.ts` / `codexTalentsStore.ts`); events come from the static JSON. `getActualExpansion`/`getActualColor` in `cardsResponseMapper.ts` run at **map time, not render time** — after changing them, force a **Resync data** (or clear `codex_cards_v2`) or you are looking at cards mapped by the old code.
+- **Card artwork is resolved through the module-scope `Map` in `@/shared/hooks/useCardImageSrc`**, never a per-render linear `.find` over the ~3100-entry `card-artwork.json` — Cardex resolves artwork once per rendered row. Lookups must use the JSON's key verbatim: some artwork names carry curly apostrophes, typos, and set suffixes (`"Typhon's Cunning"` → `"Thyphon’s cunning_eclypse-miniset"`), so no amount of transforming a card name produces the filename, and normalizing quotes on either side of the lookup breaks it. A missing name or `artwork: null` is the **only** reliable miss signal: blightbane.io serves a valid placeholder webp with HTTP 200 for non-existent icons, so `onError` never fires for a wrong artwork value.
 - Tree layout/rendering changes are verified **visually** (before/after comparison in the dev server, incl. expanded/collapsed nodes and zoom levels), not with unit tests.
